Reorganized name-ext for readability
